move "fetch", "unfetch", "add", "remove" into file
for reuse in peer for fake fetch
