+++ lib/jet/daemon.js
@@ -6,9 +6,9 @@ var EventEmitter = require('events').EventEmitter;
 var MessageSocket = require('./message-socket').MessageSocket;
 
 var jetUtils = require('./utils');
-var jetSorter = require('./daemon/sorter');
-var jetFetcher = require('./daemon/fetcher');
 var Router = require('./daemon/router').Router;
+var jetFetcher = require('./daemon/fetcher');
+var fetchCommon = require('./fetch-common');
 var JsonRPC = require('./daemon/jsonrpc');
 var Elements = require('./daemon/element').Elements;
 var Peers = require('./daemon/peers').Peers;
@@ -106,49 +106,23 @@ var createDaemon = function (options) {
 			});
 		};
 
-		var fetcher;
-		var sorter;
-		var initializing = true;
-
-		if (isDefined(params.sort)) {
-			sorter = jetSorter.create(params, queueNotification);
-			fetcher = jetFetcher.create(params, function (nparams) {
-				sorter.sorter(nparams, initializing);
-			});
-		} else {
-			fetcher = jetFetcher.create(params, queueNotification);
+		var queueSuccess = function () {
 			if (isDefined(message.id)) {
 				peer.sendMessage({
 					id: message.id,
 					result: true
 				});
 			}
-		}
+		};
 
-		peer.addFetcher(fetchId, fetcher);
-		elements.addFetcher(peer.id + fetchId, fetcher);
-		initializing = false;
-
-		if (isDefined(sorter) && sorter.flush) {
-			if (isDefined(message.id)) {
-				peer.sendMessage({
-					id: message.id,
-					result: true
-				});
-			}
-			sorter.flush();
-		}
+		fetchCommon.fetchCore(peer, elements, params, queueNotification, queueSuccess);
 	};
 
 	// dispatchers the 'unfetch' jet call.
 	// removes all ressources associated with the fetcher.
 	var unfetch = function (peer, message) {
 		var params = message.params;
-		var fetchId = checked(params, 'id', 'string');
-		var fetchPeerId = peer.id + fetchId;
-
-		peer.removeFetcher(fetchId);
-		elements.removeFetcher(fetchPeerId);
+		fetchCommon.unfetchCore(peer, elements, params);
 	};
 
 
@@ -188,15 +162,13 @@ var createDaemon = function (options) {
 	};
 
 	var add = function (peer, message) {
-		var params = message.params;
-		var path = checked(params, 'path', 'string');
-		var value = params.value;
-		elements.add(peers.eachPeerFetcher(), peer, path, value);
+		var params = checked(message, 'params', 'object');
+		fetchCommon.addCore(peer, peers.eachPeerFetcher(), elements, params);
 	};
 
 	var remove = function (peer, message) {
-		var path = checked(message.params, 'path', 'string');
-		elements.remove(path, peer);
+		var params = checked(message, 'params', 'object');
+		fetchCommon.removeCore(peer, elements, params);
 	};
 
 	var config = function (peer, message) {
